Move utils.py into package
