+++ pyethereum/utils.py
